update snapshots after refoctoring for end to end test and refactor server side tests
