+++ package.json
@@ -15,9 +15,10 @@
     "start": "node ./build/app",
     "start:dev": "nodemon --exec babel-node ./server/app",
     "start:client": "webpack-dev-server --inline --hot --config webpack.dev.js",
-    "test": "NODE_ENV=test npm run unmigrate && NODE_ENV=test npm run migrate && NODE_ENV=test nyc mocha --compilers js:babel-core/register tests/test.js",
-    "test:client": "jest --coverage",
+    "test": "NODE_ENV=test npm run unmigrate && NODE_ENV=test npm run migrate && NODE_ENV=test nyc mocha --compilers js:babel-core/register tests/index.js",
+    "test:client": "jest --coverage --silent",
     "test:watch": "npm run test:client -- --watch",
+    "test:e2e": "NODE_ENV=test npx nightwatch",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "repository": {
