+++ tests/index.js
@@ -1,0 +1,5 @@
+import './user';
+import './recipe';
+import './review';
+import './favorite';
+import './vote';
